requirements.txt
playwright
pytest
pytest-bdd
allure-pytest
pytest-rerunfailures
pytest-playwright
python-dotenv
